# By Sasebon
# Via Sasebon
* 'master' of kame.prigla.net:214b_ilohas:
  ゲーム内時間単位を秒からミリ秒に変更

+++ Classes/Scene/GameScene.h
@@ -17,7 +17,7 @@ namespace TapGun
 	{
 	public:
 
-		float timeCount;//
+		int timeCount;//
 
 		//メンバ変数
